+++ ErsatzTV.Application/Streaming/PlayoutItemProcessModel.cs
@@ -1,9 +1,11 @@
 ﻿using CliWrap;
+using ErsatzTV.Core.Interfaces.Streaming;
 
 namespace ErsatzTV.Application.Streaming;
 
 public record PlayoutItemProcessModel(
     Command Process,
+    Option<GraphicsEngineContext> GraphicsEngineContext,
     Option<TimeSpan> MaybeDuration,
     DateTimeOffset Until,
     bool IsComplete);