+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Add *experimental* graphics engine
+  - Permanent watermarks will use new graphics engine
+  - Intermittent watermarks will still use normal overlay pipeline (for now)
 
 ## [25.4.0] - 2025-08-05
 ### Added